reset pass, forget pass

+++ client/src/components/navbar/Navbar.jsx
@@ -1,13 +1,26 @@
-import React, { useState } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { clearCart, getTotals } from "../../store/cartSlice";
+import React, { useEffect, useState } from "react";
+import { useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import { sessionService } from "redux-react-session";
+import { globalStore } from "../../store/globalStore";
+import { get_user } from "../../service/decrypt";
 
 export default function Navbar() {
   const [isOpen, setIsOpen] = useState(false);
-  const dispatch = useDispatch();
   const { totalQuantity } = useSelector((state) => state.cart);
+  const [user, setUser] = useState({});
+
+  useEffect(() => {
+    if (
+      globalStore.getState().session.user !== "" &&
+      globalStore.getState().session.authenticated
+    ) {
+      const u = get_user(globalStore.getState().session.user);
+      setUser(u);
+    } else {
+      setUser({});
+    }
+  }, []);
 
   return (
     <nav className="bg-white shadow-md">
@@ -82,33 +95,44 @@ export default function Navbar() {
                 )}
               </button>
             </Link>
-            <Link to="/profile">
-              <button className="bg-gray-800 text-white px-4 py-2 rounded-md hover:bg-gray-700">
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke-width="1.5"
-                  stroke="currentColor"
-                  class="size-6"
-                >
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M17.982 18.725A7.488 7.488 0 0 0 12 15.75a7.488 7.488 0 0 0-5.982 2.975m11.963 0a9 9 0 1 0-11.963 0m11.963 0A8.966 8.966 0 0 1 12 21a8.966 8.966 0 0 1-5.982-2.275M15 9.75a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z"
-                  />
-                </svg>
-              </button>
-            </Link>
+            {user._id && (
+              <Link to="/profile">
+                <button className="bg-gray-800 text-white px-4 py-2 rounded-md hover:bg-gray-700">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke-width="1.5"
+                    stroke="currentColor"
+                    class="size-6"
+                  >
+                    <path
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      d="M17.982 18.725A7.488 7.488 0 0 0 12 15.75a7.488 7.488 0 0 0-5.982 2.975m11.963 0a9 9 0 1 0-11.963 0m11.963 0A8.966 8.966 0 0 1 12 21a8.966 8.966 0 0 1-5.982-2.275M15 9.75a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z"
+                    />
+                  </svg>
+                </button>
+              </Link>
+            )}
             {/* temp logout button */}
             <button
               className="bg-gray-800 text-white px-4 py-2 rounded-md hover:bg-gray-700"
               onClick={() => {
-                sessionService.deleteSession();
-                sessionService.deleteUser();
+                if (user._id) {
+                  sessionService.deleteSession();
+                  sessionService.deleteUser();
+                  setUser({});
+                } else {
+                  window.location.href = "/login";
+                }
               }}
             >
-              logout
+              {
+                // if user is logged in, show logout button else show login button
+
+                user._id ? "Logout" : "Login"
+              }
             </button>
           </div>
         </div>
